+++ tests/test_dataset.py
@@ -1,9 +1,14 @@
 
+from __future__ import absolute_import
+
 from nose.tools import with_setup, eq_, raises
+
 from fixture import DataSet
 from fixture.dataset import (
     Ref, DataType, DataRow, SuperSet, MergedSuperSet, is_rowlike)
-from .. import attr
+
+from . import attr
+
 
 class Books(DataSet):
     def data(self):
